Create Manage Contact class
