+++ lib/contacts.json
@@ -1,1 +1,14 @@
-[{"first_name":"Elle","last_name":"Dorie","email_address":"elle@hello.com","phone_number":"07888999000"}]
+[
+  {
+    "first_name": "Elle",
+    "last_name": "Dorie",
+    "email_address": "elle@hello.com",
+    "phone_number": "07888999000"
+  },
+  {
+    "first_name": "E",
+    "last_name": "E",
+    "email_address": "e",
+    "phone_number": ""
+  }
+]
